move functions for pep440 to `/src/lib`
