+++ src/lib/backend.ai-client-esm.ts
@@ -8,7 +8,7 @@ Licensed under MIT
 /*jshint esnext: true */
 import CryptoES from 'crypto-es';
 //var CryptoES = require("crypto-js"); /* Exclude for ES6 */
-
+import { comparePEP440Versions } from './pep440';
 type requestInfo = {
   method: string;
   headers: Headers;
@@ -680,15 +680,7 @@ class Client {
    */
   isManagerVersionCompatibleWith(version) {
     let managerVersion = this._managerVersion;
-    managerVersion = managerVersion
-      .split('.')
-      .map((s) => s.padStart(10))
-      .join('.');
-    version = version
-      .split('.')
-      .map((s) => s.padStart(10))
-      .join('.');
-    return version <= managerVersion;
+    return comparePEP440Versions(managerVersion, version) >= 0;
   }
 
   /**
